requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
Django==4.1.13
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==1.16.0
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fluidsynth==0.2
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
googletrans==4.0.2
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mongoengine==0.29.1
more-itertools==10.6.0
mpmath==1.3.0
mssql-django==1.5
networkx==3.4.2
numba==0.61.0
numpy==2.1.1
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pymongo==3.11.0
pyngrok==7.2.3
pyodbc==4.0.39
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2024.1
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
sqlparse==0.2.4
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.4
ultralytics==8.3.96
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
